# website/demo-materio-mvp/docs/openapi-gpt.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Matério — Catalogue Rénovation",
    "description": "API pour accéder au catalogue de produits Matério, centres de rénovation dans les Laurentides au Québec. Permet de rechercher des produits, analyser des projets de rénovation avec liste de matériaux complète, et obtenir les informations sur les magasins et services.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://demo-materio.vercel.app"
    }
  ],
  "paths": {
...
    "/api/gpt?action=project": {
      "post": {
        "operationId": "analyzeProject",
        "summary": "Analyser un projet de rénovation et obtenir la liste complète de matériaux",
        "description": "Analyse un projet (toiture, terrasse, sous-sol, salle de bain, cabanon, garage, clôture, peinture) et retourne tous les matériaux nécessaires avec quantités calculées, outils recommandés, et prix total. Toujours fournir les dimensions pour un calcul précis.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "action": {
                    "type": "string",
                    "enum": ["project"],
                    "default": "project"
                  },
                  "project_type": {
                    "type": "string",
                    "enum": ["toiture", "terrasse", "sous_sol", "salle_de_bain", "cabanon", "garage", "cloture", "peinture"],
                    "description": "Type de projet de rénovation ou construction"
                  },
                  "dimensions": {
                    "type": "string",
                    "description": "Dimensions du projet (ex: '1200 pi²', '12x16 pieds', '60 pieds linéaires')"
                  }
                },
                "required": ["action", "project_type"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Analyse complète du projet avec matériaux, outils et prix",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "project": {
                      "type": "object",
                      "properties": {
                        "type": { "type": "string" },
                        "name": { "type": "string" },
                        "icon": { "type": "string" },
                        "dimensions": { "type": "object" }
                      }
                    },
                    "materials": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "pid": { "type": "string" },
                          "title": { "type": "string" },
                          "price": { "type": "string" },
                          "quantity": { "type": "integer" },
                          "quantity_formula": { "type": "string" },
                          "subtotal": { "type": "string" },
                          "role": { "type": "string" }
                        }
                      }
                    },
                    "tools": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "pid": { "type": "string" },
                          "title": { "type": "string" },
                          "price": { "type": "string" },
                          "importance": { "type": "string", "enum": ["essentiel", "tres_utile", "recommande"] },
                          "explanation": { "type": "string" }
                        }
                      }
                    },
                    "pricing": {
                      "type": "object",
                      "properties": {
                        "subtotal": { "type": "string" },
                        "tps": { "type": "string" },
                        "tvq": { "type": "string" },
                        "total": { "type": "string" }
                      }
                    },
                    "financing": {
                      "type": "object",
                      "nullable": true,
                      "properties": {
                        "months": { "type": "integer" },
                        "monthly": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}